Refactor transform to polymorphic dispatch via Alg.transform_by()
Instead of a fragile isinstance chain in alg_transform.py, each Alg
subclass now implements its own transform_by(p, n_slices) method.
This ensures new Alg subtypes must handle transformation (abstract
method on Alg base class), and removes the need to access protected
members from outside.

- Added abstract transform_by() to Alg base class
- Implemented in: FaceAlg, SlicedFaceAlg, WideLayerAlg, SliceAlg,
  MiddleSliceAlg, SlicedSliceAlg, WholeCubeAlg, _Inv, _Mul, SeqAlg,
  AnnotationAlg (MarkerMeetAlg/HeadingAlg inherit)
- Extracted FacePermutation and helpers to face_permutation.py
- Simplified alg_transform.py to just compute_permutation + transform
