+++ src/cube/domain/algs/MiddleSliceAlg.py
@@ -8,7 +8,9 @@ from cube.domain.algs.SliceAlgBase import SliceAlgBase
 from cube.domain.model.cube_slice import SliceName
 
 if TYPE_CHECKING:
+    from cube.domain.algs.Alg import Alg
     from cube.domain.algs.SimpleAlg import SimpleAlg
+    from cube.domain.algs.face_permutation import FacePermutation
 
 
 @final
@@ -75,3 +77,8 @@ class MiddleSliceAlg(SliceAlgBase):
         """Return the sliceable all-slices alg as the base."""
         from cube.domain.algs.Algs import Algs
         return Algs.of_slice(self._slice_name)
+
+    def transform_by(self, p: "FacePermutation", n_slices: int | None) -> "Alg":
+        from cube.domain.algs.face_permutation import transform_slice
+        new_slice, new_n, _ = transform_slice(p, self._slice_name, self._n)
+        return MiddleSliceAlg(new_slice).with_n(new_n)
